Deprecate `src/app/core/g3w-ol/controls/utils.js`
- add new file `src/app/core/utils/ol.controls.js`
- deprecation notice for `src/app/g3w-ol/controls/utils.js`
